Parity with emscripten
For native we can assume that calls to get the adapter and the device
are synchronous, but the same isn't true for emscripten. So I've
introduced an initialisation event loop so we can wait until things are
ready before we progress any further.

Once the init_loop is complete we move onto the render loop as before.

This results in the same output for web and native, tested on Windows 10
and chrome 115.